Practised LifecycleProcessor interface:
A bean implemented LifecycleProcessor interface won't be treat as something
more than Lifecycle. That means, only the Lifecycle interface methods will 
be call in the whole life cycle, no matter whether the ctx call ctx.refresh()
or ctx.close(). For a bean itself, it should only implement the Lifecycle interface,
unless this bean is used for lifecycleProcessor within the context.