Created a project to practise 3.8 Container extension points http://static.springsource.org/spring/docs/3.0.x/spring-framework-reference/html/beans.html#beans-factory-extension
Finished 3.8.1.1 Example: Hello world, BeanPostProcessor style
It's actually a chain-of-responsibility design pattern.
The chain order is determined by the Ordered interface if provided.
The BeanPostProcessor provides to methods to process before and after 
initialisating the bean. The exact calling point of these methods are 
described as:

When such a class is registered as a post-processor with the container, 
for each bean instance that is created by the container, the post-processor 
gets a callback from the container both before container initialization 
methods (such as afterPropertiesSet and any declared init method) are called, 
and also afterwards.

Since the post-processor can handle the bean itself and can totally replace
the bean with another one or just change it's properties before and after
the bean has been initialized (using init-method), post-processor can do
anything including wraping the bean. As the document describes:

A bean post-processor typically checks for callback interfaces, or may wrap 
a bean with a proxy. Some Spring AOP infrastructure classes are implemented as 
bean post-processors and they do this proxy-wrapping logic.

The post-processor will be automatically discovered by the application context
without given a specific bean name or id. So, a typical configuration of these
post-processors are anonymous beans.

From the bean initialization point of view, BeanPostProcessor beans are special.
They will be initialized on startup, as part of the special startup phase of the 
ApplicationContext.

They will be called in order.

And they will work only within one container.